move opendata out of scaper.py
