+++ scraper/scraper.py
@@ -1,19 +1,16 @@
 import sys
-import csv
 import datetime as dt
-import io
 import json
 import os
 import traceback
 from collections import deque
 from multiprocessing import Pool
 from typing import Counter, Iterator
-from scraper.profiler import Profiling, ProfiledPool
+from scraper.profiler import Profiling
 
 from scraper.error import ScrapeError, BlockedByDoctolibError
 
 import pytz
-import requests
 
 from scraper.pattern.center_info import convert_csv_data_to_center_info, CenterInfo
 from scraper.pattern.scraper_request import ScraperRequest
@@ -29,6 +26,7 @@ from .ordoclic import centre_iterator as ordoclic_centre_iterator
 from .ordoclic import fetch_slots as ordoclic_fetch_slots
 from .mapharma.mapharma import centre_iterator as mapharma_centre_iterator
 from .mapharma.mapharma import fetch_slots as mapharma_fetch_slots
+from .opendata.opendata import center_iterator as gouv_centre_iterator
 from random import random
 
 POOL_SIZE = int(os.getenv("POOL_SIZE", 50))
@@ -283,52 +281,6 @@ def centre_iterator():  # pragma: no cover
         if center["rdv_site_web"] not in visited_centers_links:
             visited_centers_links.add(center["rdv_site_web"])
             yield center
-
-
-def gouv_centre_iterator(outpath_format="data/output/{}.json"):
-    url = "https://www.data.gouv.fr/fr/datasets/r/5cb21a85-b0b0-4a65-a249-806a040ec372"
-    response = requests.get(url)
-    response.raise_for_status()
-
-    reader = io.StringIO(response.content.decode("utf8"))
-    csvreader = csv.DictReader(reader, delimiter=";")
-
-    total = 0
-
-    centres_non_pris_en_compte = {"centres_fermes": {}, "centres_urls_vides": []}
-
-    for row in csvreader:
-
-        row["rdv_site_web"] = fix_scrap_urls(row["rdv_site_web"])
-        if row["centre_fermeture"] == "t":
-            centres_non_pris_en_compte["centres_fermes"][row["gid"]] = row["rdv_site_web"]
-        if should_use_opendata_csv(row["rdv_site_web"]):
-            yield row
-        else:
-            centres_non_pris_en_compte["centres_urls_vides"].append(row["gid"])
-
-        total += 1
-
-    nb_fermes = len(centres_non_pris_en_compte["centres_fermes"])
-    nb_urls_vides = len(centres_non_pris_en_compte["centres_urls_vides"])
-
-    logger.info(f"Il y a {nb_fermes} centres fermes dans le fichier gouv sur un total de {total}")
-
-    nb_urls_vides = len(centres_non_pris_en_compte["centres_urls_vides"])
-    logger.info(f"Il y a {nb_urls_vides} centres avec une URL vide dans le fichier gouv sur un total de {total}")
-
-    outpath = outpath_format.format("centres_non_pris_en_compte_gouv")
-    with open(outpath, "w") as fichier:
-        json.dump(centres_non_pris_en_compte, fichier, indent=2)
-
-
-def should_use_opendata_csv(rdv_site_web: str) -> bool:
-    plateformes_hors_csv = ["doctolib", "maiia"]
-
-    if any(p in rdv_site_web for p in plateformes_hors_csv):
-        return False
-    return True
-
 
 def copy_omit_keys(d, omit_keys):
     return {k: d[k] for k in set(list(d.keys())) - set(omit_keys)}
